requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.2
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorlog==6.7.0
contourpy==1.0.5
cowsay==4.0
cycler==0.11.0
emoji==1.7.0
fonttools==4.33.3
fortune==1.1.0
frozenlist==1.3.0
grizzled-python==2.2.0
hikari==2.0.0.dev115
hikari-lightbulb==2.3.1
idna==3.3
kiwisolver==1.4.3
langcodes==3.3.0
language-data==1.1
marisa-trie==0.7.7
matplotlib==3.6.0rc2
multidict==6.0.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
regex==2022.3.15
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
uvloop==0.16.0; sys_platform != "win32"
yarl==1.7.2
humanize==4.3.0
wordcloud==1.8.2.2
pilmoji==2.0.2
imagehash==4.3.1
pyyaml==6.0.1
hikari-toolbox==0.1.4
